feat: cursor repository
