+++ actions-indexer-repository/src/errors/cursor_repository.rs
@@ -1,0 +1,11 @@
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+/// Represents errors that can occur within the cursor repository.
+///
+/// This enum consolidates various error conditions specific to database interactions,
+/// such as SQLx errors during database operations.
+pub enum CursorRepositoryError {
+    #[error("Database error: {0}")]
+    DatabaseError(#[from] sqlx::Error),
+}
